refacto: create a new dir containing Features and al
Before this commit, `Feature`, `FeatureCollection` and `FeatureGeometry`
were all in the same file. Now they are divided in a new folder,
`src/Core/Objects`, along with `Style` which is a basic object that
should go along the precedent ones.

As it could be better to have a more understandable core for external
user, for example when writing a parser for a format not yet supported,
I wish to simplify the current API on these classes. This commit is a
first step to this goal.